refactor: move packages out of the core package
1. Move the data package to top level
2. Move the core.tracking package to data.tracking package
3. Move the core.tracking.service to background.tracking.service package
