--- app/src/main/java/com/sdevprem/runtrack/core/data/model/Gender.kt
@@ -1,5 +1,0 @@
-package com.sdevprem.runtrack.core.data.model
-
-enum class Gender {
-    MALE, FEMALE
-}
