--- app/src/main/java/com/sdevprem/runtrack/core/tracking/notification/TrackingNotificationHelper.kt
@@ -1,15 +1,0 @@
-package com.sdevprem.runtrack.core.tracking.notification
-
-import androidx.core.app.NotificationCompat
-
-interface TrackingNotificationHelper {
-    val baseNotificationBuilder: NotificationCompat.Builder
-
-    fun createNotificationChannel()
-    fun updateTrackingNotification(durationInMillis: Long, isTracking: Boolean)
-    fun removeTrackingNotification()
-
-    companion object {
-        const val TRACKING_NOTIFICATION_ID = 3
-    }
-}
