+++ app/src/main/java/com/sdevprem/runtrack/data/model/Gender.kt
@@ -1,0 +1,5 @@
+package com.sdevprem.runtrack.data.model
+
+enum class Gender {
+    MALE, FEMALE
+}
