+++ app/src/main/java/com/sdevprem/runtrack/data/tracking/location/DefaultLocationTrackingManager.kt
@@ -1,4 +1,4 @@
-package com.sdevprem.runtrack.core.tracking.location
+package com.sdevprem.runtrack.data.tracking.location
 
 import android.annotation.SuppressLint
 import android.content.Context
