+++ app/src/main/java/com/sdevprem/runtrack/domain/usecase/GetCurrentRunStateWithCaloriesUseCase.kt
@@ -1,7 +1,7 @@
 package com.sdevprem.runtrack.domain.usecase
 
 import com.sdevprem.runtrack.common.utils.RunUtils
-import com.sdevprem.runtrack.core.data.repository.UserRepository
+import com.sdevprem.runtrack.data.repository.UserRepository
 import com.sdevprem.runtrack.domain.model.CurrentRunStateWithCalories
 import com.sdevprem.runtrack.domain.tracking.TrackingManager
 import kotlinx.coroutines.flow.Flow
