+++ app/src/main/java/com/sdevprem/runtrack/ui/screen/home/HomeScreenState.kt
@@ -1,7 +1,7 @@
 package com.sdevprem.runtrack.ui.screen.home
 
-import com.sdevprem.runtrack.core.data.model.Run
-import com.sdevprem.runtrack.core.data.model.User
+import com.sdevprem.runtrack.data.model.Run
+import com.sdevprem.runtrack.data.model.User
 import com.sdevprem.runtrack.domain.model.CurrentRunStateWithCalories
 
 data class HomeScreenState(
