+++ app/src/main/java/com/sdevprem/runtrack/ui/screen/home/HomeViewModel.kt
@@ -4,9 +4,9 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.sdevprem.runtrack.common.extension.setDateToWeekFirstDay
 import com.sdevprem.runtrack.common.extension.setDateToWeekLastDay
-import com.sdevprem.runtrack.core.data.model.Run
-import com.sdevprem.runtrack.core.data.repository.AppRepository
-import com.sdevprem.runtrack.core.data.repository.UserRepository
+import com.sdevprem.runtrack.data.model.Run
+import com.sdevprem.runtrack.data.repository.AppRepository
+import com.sdevprem.runtrack.data.repository.UserRepository
 import com.sdevprem.runtrack.di.ApplicationScope
 import com.sdevprem.runtrack.di.IoDispatcher
 import com.sdevprem.runtrack.domain.tracking.TrackingManager
