+++ app/src/main/java/com/sdevprem/runtrack/ui/screen/profile/ProfileViewModel.kt
@@ -3,8 +3,8 @@ package com.sdevprem.runtrack.ui.screen.profile
 import android.net.Uri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.sdevprem.runtrack.core.data.repository.AppRepository
-import com.sdevprem.runtrack.core.data.repository.UserRepository
+import com.sdevprem.runtrack.data.repository.AppRepository
+import com.sdevprem.runtrack.data.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
