+++ app/src/main/java/com/sdevprem/runtrack/ui/screen/profile/TopBarProfile.kt
@@ -46,7 +46,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.sdevprem.runtrack.R
-import com.sdevprem.runtrack.core.data.model.User
+import com.sdevprem.runtrack.data.model.User
 import com.sdevprem.runtrack.ui.common.compose.component.UserProfilePic
 import com.sdevprem.runtrack.ui.common.extension.bottomBorder
