+++ app/src/main/java/com/sdevprem/runtrack/ui/screen/runninghistory/RunningHistoryScreen.kt
@@ -43,8 +43,8 @@ import androidx.paging.LoadState
 import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
 import com.sdevprem.runtrack.R
-import com.sdevprem.runtrack.core.data.model.Run
-import com.sdevprem.runtrack.core.data.utils.RunSortOrder
+import com.sdevprem.runtrack.data.model.Run
+import com.sdevprem.runtrack.data.utils.RunSortOrder
 import com.sdevprem.runtrack.ui.common.compose.component.DropDownList
 import com.sdevprem.runtrack.ui.common.compose.component.RunInfoDialog
 import com.sdevprem.runtrack.ui.common.compose.component.RunItem
