+++ app/src/main/java/com/sdevprem/runtrack/ui/screen/runstats/RunStatsViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.sdevprem.runtrack.common.extension.setDateToWeekFirstDay
 import com.sdevprem.runtrack.common.extension.setDateToWeekLastDay
-import com.sdevprem.runtrack.core.data.repository.AppRepository
+import com.sdevprem.runtrack.data.repository.AppRepository
 import com.sdevprem.runtrack.di.DefaultDispatcher
 import com.sdevprem.runtrack.ui.screen.runstats.utils.RunStatsAccumulator
 import dagger.hilt.android.lifecycle.HiltViewModel
